Created a package of the cnn network
